requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
anyio==4.1.0
    # via jupyter-server
appnope==0.1.3
    # via ipykernel
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==2.4.1
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.13.1
    # via jupyterlab-server
beautifulsoup4==4.12.2
    # via nbconvert
bleach==6.1.0
    # via nbconvert
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via argon2-cffi-bindings
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
comm==0.2.0
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
exceptiongroup==1.2.0
    # via
    #   anyio
    #   ipython
executing==2.0.1
    # via stack-data
fastjsonschema==2.19.0
    # via nbformat
filelock==3.13.1
    # via torch
fonttools==4.46.0
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
fsspec==2023.12.0
    # via torch
idna==3.6
    # via
    #   anyio
    #   jsonschema
    #   requests
importlib-metadata==6.9.0
    # via
    #   build
    #   jupyter-client
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
importlib-resources==6.1.1
    # via matplotlib
ipycanvas==0.13.1
    # via -r requirements.in
ipykernel==6.27.1
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
    #   qtconsole
ipython==8.18.1
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==8.1.1
    # via
    #   ipycanvas
    #   jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.1
    # via ipython
jinja2==3.1.2
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   torch
json5==0.9.14
    # via jupyterlab-server
jsonpointer==2.4
    # via jsonschema
jsonschema[format-nongpl]==4.20.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.11.2
    # via jsonschema
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==8.6.0
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
    #   qtconsole
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.5.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
    #   qtconsole
jupyter-events==0.9.0
    # via jupyter-server
jupyter-lsp==2.2.1
    # via jupyterlab
jupyter-server==2.11.1
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.4.4
    # via jupyter-server
jupyterlab==4.0.9
    # via notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.25.2
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.9
    # via ipywidgets
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.3
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.8.2
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==3.0.2
    # via nbconvert
mpmath==1.3.0
    # via sympy
nbclient==0.9.0
    # via nbconvert
nbconvert==7.11.0
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.9.2
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.8
    # via ipykernel
networkx==3.2.1
    # via torch
notebook==7.0.6
    # via jupyter
notebook-shim==0.2.3
    # via
    #   jupyterlab
    #   notebook
numpy==1.26.2
    # via
    #   contourpy
    #   ipycanvas
    #   matplotlib
    #   pandas
overrides==7.4.0
    # via jupyter-server
packaging==23.2
    # via
    #   build
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
    #   qtconsole
    #   qtpy
pandas==2.1.4
    # via -r requirements.in
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.1.0
    # via
    #   -r requirements.in
    #   ipycanvas
    #   matplotlib
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.0.0
    # via jupyter-core
prometheus-client==0.19.0
    # via jupyter-server
prompt-toolkit==3.0.41
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.6
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyparsing==3.1.1
    # via matplotlib
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
    #   pandas
python-json-logger==2.0.7
    # via jupyter-events
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via jupyter-events
pyzmq==25.1.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   qtconsole
qtconsole==5.5.1
    # via jupyter
qtpy==2.4.1
    # via qtconsole
referencing==0.31.1
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.31.0
    # via jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.13.2
    # via
    #   jsonschema
    #   referencing
send2trash==1.8.2
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.0
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
sympy==1.12
    # via torch
terminado==0.18.0
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.2.1
    # via nbconvert
tomli==2.0.1
    # via
    #   build
    #   jupyterlab
    #   pip-tools
    #   pyproject-hooks
torch==2.1.1
    # via -r requirements.in
tornado==6.4
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
tqdm==4.66.1
    # via -r requirements.in
traitlets==5.14.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   qtconsole
types-python-dateutil==2.8.19.14
    # via arrow
typing-extensions==4.8.0
    # via
    #   async-lru
    #   ipython
    #   torch
tzdata==2023.4
    # via pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.1.0
    # via requests
wcwidth==0.2.12
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.6.4
    # via jupyter-server
wheel==0.42.0
    # via pip-tools
widgetsnbextension==4.0.9
    # via ipywidgets
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
